+++ scrapper/src/main/resources/db/changelog/changeset/create-tag-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE tags (
-    id SERIAL PRIMARY KEY,
-    link_id INTEGER NOT NULL,
+    id BIGSERIAL PRIMARY KEY,
+    link_id BIGINT NOT NULL,
     tag TEXT NOT NULL,
     FOREIGN KEY (link_id) REFERENCES links(id) ON DELETE CASCADE,
     UNIQUE (link_id, tag)
